feat: make cart-types for type TCartItem
