+++ src/types/cart-types.ts
@@ -1,0 +1,8 @@
+export type TCartItem = {
+  id: string
+  name: string
+  image: string
+  price: number
+  quantity: number
+  total: number
+}
