+++ src/types/user-types.ts
@@ -20,12 +20,3 @@ export type TCustomerProfileForm<T> = {
   defaultShipping: T
   defaultBilling: T
 }
-
-export type TCartItem = {
-  id: string
-  name: string
-  image: string
-  price: number
-  quantity: number
-  total: number
-}
